Fix FTBFS when using debhelper 7 and not building the lib32 packages.
The reason for depending on debhelper 8 was because it was IMO the cleanest/best
solution for this. Since the dependency got downgraded in a previous commit then
I'm including the "fix" I used before I upgraded the dependency. I could also
have done dh $@ -s but it was overkill IMO. This issue mainly affects distros
like Ubuntu Lucid that still ship with debhelper 7. This commits makes the
package compile when using debhelper 7 or the smarter debhelper 8.